requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.0.1
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
ctransformers==0.2.27
llama-cpp-python==0.2.11
huggingface-hub==0.19.4
sentence-transformers==2.2.2 